requirements.txt
requests~=2.32.4
beautifulsoup4~=4.13.4
python-dotenv~=1.1.1
pytest~=8.4.1